Implement authorization using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -64,6 +64,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, 'Access Denied');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -76,6 +80,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, 'Access Denied');
+        }
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'You question has been updated.');
     }
@@ -88,6 +96,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question))
+        {
+            abort(403, 'Access Denied');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted.');
     }
